+++ internal/service/privatelinkendpointserverless/resource_privatelink_endpoint_serverless_test.go
@@ -1,4 +1,4 @@
-package privatelinkendpointserviceserverless_test
+package privatelinkendpointserverless_test
 
 import (
 	"context"
