+++ mongodbatlas/resource_serverless_instance.go
@@ -20,8 +20,7 @@ import (
 )
 
 const (
-	errorServerlessInstanceListStatus = "error awaiting serverless instance list status IDLE: %s"
-	errorServerlessInstanceSetting    = "error setting `%s` for serverless instance (%s): %s"
+	errorServerlessInstanceSetting = "error setting `%s` for serverless instance (%s): %s"
 )
 
 func ResourceServerlessInstance() *schema.Resource {
@@ -387,36 +386,6 @@ func resourceServerlessInstanceRefreshFunc(ctx context.Context, name, projectID 
 		}
 
 		return c, c.StateName, nil
-	}
-}
-
-func resourceServerlessInstanceListRefreshFunc(ctx context.Context, projectID string, client *matlas.Client) retry.StateRefreshFunc {
-	return func() (any, string, error) {
-		c, resp, err := client.ServerlessInstances.List(ctx, projectID, nil)
-
-		if err != nil && strings.Contains(err.Error(), "reset by peer") {
-			return nil, "REPEATING", nil
-		}
-
-		if err != nil && c == nil && resp == nil {
-			return nil, "", err
-		} else if err != nil {
-			if resp.StatusCode == 404 {
-				return "", "DELETED", nil
-			}
-			if resp.StatusCode == 503 {
-				return "", "PENDING", nil
-			}
-			return nil, "", err
-		}
-
-		for i := range c.Results {
-			if c.Results[i].StateName != "IDLE" {
-				return c.Results[i], "PENDING", nil
-			}
-		}
-
-		return c, "IDLE", nil
 	}
 }
